Switch elb/ec2 over to signature version 4
This matches the public documentation and other AWS SDKs, which
all use signature version 4 for these services.

Note that this is no longer a detect_potential_sigv4, we should be
using signature version 4 at all times for ec2/elb.

I've also updated an assertion method to not require that the
ignore_params_values actually exist in the request.  We made
a similar change in the AWS CLI a while back.  All we care about is
that we're not validating those params.  We don't actually care
if they're in the serialized params are not.